Move upgrade_hostos to hostos_upgrade.rs
